Add files via upload
In today's world, many organisations and institutions are facing light-related issues and are seeking automated or smart systems to address these issues. The purpose of such systems is to effectively manage and control lighting in order to save energy and optimise its usage. By automating light control, wastage can be minimised and optimal utilisation can be achieved.

This proposed system is based on the detection of humans using the Yolov3 deep learning algorithm. When a human presence is detected, the system will send notifications via email and text messages. By implementing this system, both the management and employees (such as teachers and staff) of the organisation or institute will find it more feasible to have the lighting automatically controlled when there are no individuals present, thereby reducing unnecessary energy consumption.

What makes this system particularly advantageous is that it can be implemented without the need for additional automatic sensors. It leverages the existing CCTV cameras or webcams to detect human presence and control the lighting accordingly.